backend/requirements.txt
django
faker
djangorestframework
djangorestframework-simplejwt
django-cors-headers